cephadm: move snmp class to a new file
